--- src/main/java/com/example/handlemanager/HandleManagerApplication.java
@@ -1,16 +1,0 @@
-package com.example.handlemanager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-
-
-@SpringBootApplication
-@EnableCaching
-public class HandleManagerApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(HandleManagerApplication.class, args);
-  }
-
-}
