--- src/main/java/com/example/handlemanager/domain/model/PidStatus.java
@@ -1,12 +1,0 @@
-package com.example.handlemanager.domain.model;
-
-public enum PidStatus {
-  DRAFT,
-  ACTIVE,
-  ARCHIVED,
-  DEPRECIATED,
-  FAILED,
-  MERGED,
-  SPLIT,
-  TEST
-}
