--- src/main/java/com/example/handlemanager/domain/responses/DoiRecordResponse.java
@@ -1,28 +1,0 @@
-package com.example.handlemanager.domain.responses;
-
-import com.example.handlemanager.repositoryobjects.Handles;
-import java.util.List;
-import lombok.Getter;
-
-@Getter
-public class DoiRecordResponse extends HandleRecordResponse {
-
-  String referentDoiName;
-  String referent;
-
-  public DoiRecordResponse(List<Handles> entries) {
-    super(entries);
-
-    String type;
-
-    for (Handles h : entries) {
-      type = h.getType();
-      if (type.equals("referentDoiName")) {
-        this.referentDoiName = h.getData();
-      }
-      if (type.equals("referent")) {
-        this.referent = h.getData();
-      }
-    }
-  }
-}
