--- src/main/java/com/example/handlemanager/exceptions/PidCreationException.java
@@ -1,9 +1,0 @@
-package com.example.handlemanager.exceptions;
-
-public class PidCreationException extends Exception {
-
-  public PidCreationException(String s) {
-    super(s);
-  }
-
-}
