--- src/main/java/com/example/handlemanager/exceptions/PidResolutionException.java
@@ -1,9 +1,0 @@
-package com.example.handlemanager.exceptions;
-
-public class PidResolutionException extends Exception {
-
-  public PidResolutionException(String s) {
-    super(s);
-  }
-
-}
