+++ src/main/java/eu/dissco/core/handlemanager/domain/model/PidStatus.java
@@ -1,0 +1,12 @@
+package eu.dissco.core.handlemanager.domain.model;
+
+public enum PidStatus {
+  DRAFT,
+  ACTIVE,
+  ARCHIVED,
+  DEPRECIATED,
+  FAILED,
+  MERGED,
+  SPLIT,
+  TEST
+}
