+++ src/main/java/eu/dissco/core/handlemanager/domain/requests/DigitalSpecimenRequest.java
@@ -1,0 +1,59 @@
+package eu.dissco.core.handlemanager.domain.requests;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.util.Objects;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+@JsonIgnoreProperties
+public class DigitalSpecimenRequest extends DoiRecordRequest {
+
+  protected final String digitalOrPhysical;
+  protected final String specimenHostPid;
+  protected final String inCollectionFacilityPid;
+
+  public DigitalSpecimenRequest(
+      // Handle
+      String pidIssuerPid,
+      String digitalObjectTypePid,
+      String digitalObjectSubtypePid,
+      String[] locations,
+      // Doi
+      String referentDoiName,
+      // Digital Specimen
+      String digitalOrPhysical,
+      String specimenHostPid,
+      String inCollectionFacilityPid
+  ) {
+    super(pidIssuerPid, digitalObjectTypePid, digitalObjectSubtypePid, locations, referentDoiName);
+    this.digitalOrPhysical = digitalOrPhysical;
+    this.specimenHostPid = specimenHostPid;
+    this.inCollectionFacilityPid = inCollectionFacilityPid;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    if (!super.equals(o)) {
+      return false;
+    }
+    DigitalSpecimenRequest that = (DigitalSpecimenRequest) o;
+    return getDigitalOrPhysical().equals(that.getDigitalOrPhysical())
+        && getSpecimenHostPid().equals(
+        that.getSpecimenHostPid()) && getInCollectionFacilityPid().equals(
+        that.getInCollectionFacilityPid());
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(super.hashCode(), getDigitalOrPhysical(), getSpecimenHostPid(),
+        getInCollectionFacilityPid());
+  }
+}
