+++ src/main/java/eu/dissco/core/handlemanager/domain/responses/DoiRecordResponse.java
@@ -1,0 +1,28 @@
+package eu.dissco.core.handlemanager.domain.responses;
+
+import eu.dissco.core.handlemanager.repositoryobjects.Handles;
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+public class DoiRecordResponse extends HandleRecordResponse {
+
+  String referentDoiName;
+  String referent;
+
+  public DoiRecordResponse(List<Handles> entries) {
+    super(entries);
+
+    String type;
+
+    for (Handles h : entries) {
+      type = h.getType();
+      if (type.equals("referentDoiName")) {
+        this.referentDoiName = h.getData();
+      }
+      if (type.equals("referent")) {
+        this.referent = h.getData();
+      }
+    }
+  }
+}
