+++ src/main/java/eu/dissco/core/handlemanager/exceptions/PidCreationException.java
@@ -1,0 +1,9 @@
+package eu.dissco.core.handlemanager.exceptions;
+
+public class PidCreationException extends Exception {
+
+  public PidCreationException(String s) {
+    super(s);
+  }
+
+}
