+++ src/main/java/eu/dissco/core/handlemanager/exceptions/PidResolutionException.java
@@ -1,0 +1,9 @@
+package eu.dissco.core.handlemanager.exceptions;
+
+public class PidResolutionException extends Exception {
+
+  public PidResolutionException(String s) {
+    super(s);
+  }
+
+}
