+++ src/main/java/eu/dissco/core/handlemanager/utils/Resources.java
@@ -1,0 +1,90 @@
+package eu.dissco.core.handlemanager.utils;
+
+import eu.dissco.core.handlemanager.repositoryobjects.Handles;
+import java.io.StringWriter;
+import java.nio.charset.StandardCharsets;
+import java.util.List;
+import javax.xml.XMLConstants;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.OutputKeys;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
+import org.w3c.dom.Document;
+
+public class Resources {
+
+  private Resources() {
+    throw new IllegalStateException("Utility class");
+  }
+
+  public static String getDataFromType(String type, List<Handles> hList) {
+    for (Handles h : hList) {
+      if (h.getType().equals(type)) {
+        return h.getData();
+      }
+    }
+    return ""; // This should maybe return a warning?
+  }
+
+  public static Handles genAdminHandle(byte[] handle, long timestamp) {
+    return new Handles(handle, 100, "HS_ADMIN".getBytes(), decodeAdmin(), timestamp);
+  }
+
+  private static byte[] decodeAdmin() {
+    String admin = "0fff000000153330303a302e4e412f32302e353030302e31303235000000c8";
+    byte[] adminByte = new byte[admin.length() / 2];
+    for (int i = 0; i < admin.length(); i += 2) {
+      adminByte[i / 2] = hexToByte(admin.substring(i, i + 2));
+    }
+    return adminByte;
+  }
+
+  private static byte hexToByte(String hexString) {
+    int firstDigit = toDigit(hexString.charAt(0));
+    int secondDigit = toDigit(hexString.charAt(1));
+    return (byte) ((firstDigit << 4) + secondDigit);
+  }
+
+  private static int toDigit(char hexChar) {
+    return Character.digit(hexChar, 16);
+  }
+
+  public static byte[] setLocations(String[] objectLocations)
+      throws TransformerException, ParserConfigurationException {
+    DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
+    dbf.setFeature("http://apache.org/xml/features/disallow-doctype-decl", true);
+
+    DocumentBuilder documentBuilder = dbf.newDocumentBuilder();
+
+    var doc = documentBuilder.newDocument();
+    var locations = doc.createElement("locations");
+    doc.appendChild(locations);
+    for (int i = 0; i < objectLocations.length; i++) {
+
+      var locs = doc.createElement("location");
+      locs.setAttribute("id", String.valueOf(i));
+      locs.setAttribute("href", objectLocations[i]);
+      locs.setAttribute("weight", "0");
+      locations.appendChild(locs);
+    }
+    return documentToString(doc).getBytes(StandardCharsets.UTF_8);
+  }
+
+  private static String documentToString(Document document) throws TransformerException {
+    TransformerFactory tf = TransformerFactory.newInstance();
+    tf.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+    tf.setAttribute(XMLConstants.ACCESS_EXTERNAL_DTD, "");
+    tf.setAttribute(XMLConstants.ACCESS_EXTERNAL_STYLESHEET, "");
+
+    var transformer = tf.newTransformer();
+    transformer.setOutputProperty(OutputKeys.OMIT_XML_DECLARATION, "yes");
+    StringWriter writer = new StringWriter();
+    transformer.transform(new DOMSource(document), new StreamResult(writer));
+    return writer.getBuffer().toString();
+  }
+
+}
